LibWeb: Move ComponentValue code into ComponentValue.cpp
