+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -44,6 +44,7 @@ set(SOURCES
     CSS/MediaList.cpp
     CSS/MediaQuery.cpp
     CSS/MediaQueryList.cpp
+    CSS/Parser/ComponentValue.cpp
     CSS/Parser/Parser.cpp
     CSS/Parser/StyleRules.cpp
     CSS/Parser/Token.cpp
